Prompt notice when tor is starting
When the prompt script is first ran it can take a while for tor to start.
Giving a notice that tor is starting.

This also changes the Controller instance variable from 'control' to
'controller'. I'm trying to standardize on the later to avoid a naming confict
with stem.control imports.