log_to_stdout() only printed at TRACE level
Ok, the logging module is *trying* to make me hate it. I haven't a clue what's
wrong with the basicConfig() function, but when I try to use it the 'level'
argument is ignored and it always spews its output at the TRACE runlevel. This
is worthless (trace is often too noisy to be helpful) so making my own
StdoutLogger.

Tested by running this on a separate system, then copying the change over by
hand.