Remove extra checks for circuit id and stream id validity
CircuitID and StreamID may not be None, so raise ProtocolError if either
ID is None.  Inside is_valid_circuit_id() and is_valid_stream_id(),
isinstance(entry, str) already returns False if None is checked.  Add None
as invalid IDs in tests.

NOTE: Do not try this with is_valid_fingerprint() and is_valid_nickname(),
as the fingerprint or nickname may be None in certain cases.

Signed-off-by: Sean Robinson <seankrobinson@gmail.com>