Dropping duplicate stop() call
The test runner is ran within a finally block that calls() stop. Reguardless of
if tor fails to start or not stop() will be called, so this call was a
duplicate.

Calling stop() multiple times is fine (the method is idempotent), but it logs
each time so we were saying "Shutting down tor... done" twice. It's still
possible for us to give double messages in some other use cases so we should
address this more generally later, though I'm not sure offhand what that fix
should be.