Moving the ExitPolicyRule's masked_bits behind a method
In looking at TorCtl's ExitPolicyLine class I'm realizing that we may want to
use our mask attribute as our main subnet representation rather than the
masked_bits.

Changing the class so the masked_bits are accessed via a method so we'll have
the flexability to change this in the future.