Allow for blocking circuit build and extension
Callers of new_circuit() and extend_circuit() will often (maybe usually?) want
their call to block until the circuit is actually ready to be used.

Unfortunately this is a little tricky for them to hack together on their own
since it requires a listener (polling doesn't work reliably - after a couple of
hours of hair pulling I now know that all too well). Hence doing this work for
them.

This includes a CircuitExtensionFailed exception so we can communicate the
circuit event that reported the failure.