Using a lightweight rule counterpart for microdescriptor policies
The exit policies used in microdescriptors are wonderfully simple. There's no
point in storing full ExitPolicyRule instances for them, so making an alternate
lightweight subclass.

Advantages are faster for parsing and less memory usage. Disadvantage is that
this doesn't use the ExitPolicyRule constructor, so we'll need to take care
that ExitPolicyRule changes are reflected in this subclass.

This dropped the runtime of the network status integ tests on my system from
9.4 to 8.3 seconds.