Dropping the Config's synchronize() method
Choice is highly overrated. Actually, in the world of libraries providing too
many options of doing a single can make things damn confusing.

The Config's synchronize() method was a great improvement in its day, but now
that we have config_dict() it's redundant. Hopefully our module's docs are now
more friendly for newcomers.