Running PEP8 checks under a '--style' arg
Well... damn. I had hoped that we would be able to wrap the PEP8 style checks
into our normal testing runs (the reason check_whitespace.py works so well
is that it's trivial to do with every test run).

However, pep8 takes on the order of ten seconds to run over our codebase. This
is twice as long as all of our unit tests and hence too long to include in
tests without opting in.

I'm keeping my fingers crossed that they're doing something stupid with ignored
issues so we'll get better performance when we've made it happier. We'll see.

On the up side PEP8 is wonderfully configurable with respect to what it accepts
and ignores so we can start using it right away. Also, the output is well
formed so it was trivial to wrap it into our present style checker (which means
consolidated and colorized output - yay!).